Move PropertyTypeRegistry to org.eclipse.ec4j.core because it is not a model class
